add auth middleware

+++ db/query/account.sql
@@ -17,9 +17,10 @@ FOR NO KEY UPDATE;
 
 -- name: ListAccounts :many
 SELECT * FROM accounts
+WHERE owner = $1
 ORDER BY id
-LIMIT $1
-OFFSET $2;
+LIMIT $2
+OFFSET $3;
 
 -- name: UpdateAccount :one
 UPDATE accounts
@@ -33,5 +34,5 @@ SET balance = balance + sqlc.arg(amount)
 WHERE id = sqlc.arg(id)
 RETURNING *;
 
--- name: DeleteAccount :exec
-DELETE FROM accounts WHERE id = $1;
+-- name: DeleteAccount :execrows
+DELETE FROM accounts WHERE id = $1 and owner = $2;
